requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
cfgv==3.4.0
click==8.1.7
distlib==0.3.7
fastapi==0.103.2
filelock==3.13.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.0
identify==2.5.31
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
nodeenv==1.8.0
platformdirs==3.11.0
pre-commit==3.5.0
psycopg2==2.9.8
psycopg2-binary==2.9.8
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
virtualenv==20.24.6
watchfiles==0.20.0
websockets==11.0.3
